Cleanup: propoerly inject things to PushRulesFragment and move PushRulesController to its own file
